requirements.txt
transformers >= 4.39.3
bertopic >= 0.16.0
keybert >= 0.8.4
rake-nltk >= 1.0.6
datasets >= 2.18.0
flask >= 2.2.2
flask-cors >= 4.0.0
werkzeug >= 2.2.2
pandas >= 1.5.3
sqlalchemy >= 1.4.39
mysql-connector-python >= 8.3.0
app_store_scraper >= 0.3.5
google_play_scraper >= 1.2.6
numpy >= 1.23.5
scikit-learn >= 1.2.1
huggingface_hub >= 0.22.2
accelerate >= 0.29.3
seaborn >= 0.12.2
tqdm >= 4.66.2
